server/requirements.txt
Flask
flask-sqlalchemy
flask-migrate
psycopg2-binary
flask-cors
werkzeug
python-dotenv